+++ modules/canbus/vehicle/gem/protocol/wiper_cmd_90.cc
@@ -61,7 +61,7 @@ void Wipercmd90::set_p_wiper_cmd(uint8_t* data,
   uint8_t x = wiper_cmd;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 0, 8);
+  to_set.set_value(static_cast<uint8_t>(x), 0, 8);
 }
 
 }  // namespace gem
